remove optimism sdk
