+++ packages/params/src/gas-estimation/index.ts
@@ -1,4 +1,3 @@
 export * from "./arbitrum";
 export * from "./optimism";
 export * from "./mantle";
-export * from "./ancient8";
